[Bronze V] Title: A+B - 9, Time: 0 ms, Memory: 2020 KB, Score: 75 point -BaekjoonHub

+++ 백준/Bronze/15740. A＋B － 9/A＋B － 9.cc
@@ -1,7 +1,7 @@
 #include<iostream>
 
 int main() {
-    int a, b =0;
+    long long a, b =0;
     std::cin >> a;
     std::cin >> b;
     std::cout << a+b;
